pyproject.toml
[project]
name = "shelf-occupancy-analyzer"
version = "0.1.0"
description = "Sistema de análisis de ocupación de anaqueles usando estimación de profundidad"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "scikit-image>=0.21.0",
    "scipy>=1.11.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tqdm>=4.65.0",
    "loguru>=0.7.0",
    "huggingface-hub>=0.19.0",
    "requests>=2.31.0",
    "scikit-learn>=1.7.2",
    "ipykernel>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.287",
    "mypy>=1.5.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shelf_occupancy"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
